# swagger/swagger.yaml
swagger: '2.1'
info:
  title: Docker Vulnerability Manager API
  description: Docker containers' source code freshness management API
  version: 0.19.0
consumes:
  - application/json
produces:
  - application/json
basePath: /api
components:
  parameters:
    token:
      in: header
      name: token
      required: true
      schema:
        type: string
      description: authentication token
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVhOTVhNzBiYjljMjQ1MzA2YjE2YzE0NSIsImlhdCI6MTUxOTkyMzc5NiwiZXhwIjoxNTE5OTI3Mzk2fQ.e5cKEEGtDb54AVlnZHf-P132x1LTrS4WrgXywfmkSkk
...
    imageName:
      in: body
      name: imageName
      required: true
      schema:
        type: string
      description: name of the Docker image
      example: pawelpaszki/vuln-demo-2-node
...
    dockerUsername:
      in: body
      name: username
      required: true
      schema:
        type: string
      descrpiton: Docker registry username
    dockerPassword:
      in: body
      name: password
      required: true
      schema:
        type: string
      descrpiton: Docker registry password
...
paths:
...
  /src/availableDirs:
    get:
      summary: extracted images
      description: get a list of names of images with extracted source code
      operationId: getAvailableDirs
      tags:
        - Src
      parameters:
        - $ref: '#/components/parameters/token'
      responses:
        '200':
          description: OK
          schema:
            type: array
  /dockerLogin:
    post:
      summary: Docker login
      description: login to Docker registry
      operationId: dockerLogin
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/dockerUsername'
        - $ref: '#/components/parameters/dockerPassword'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
                example: Login Successful
        '401':
          description: Unable to create image freshness entry
          schema:
            type: object
            properties:
              message:
                type: string
                example: Incorrect login and/or password
  /src/checkOS:
    post:
      summary: Check OS
      description: check operating system of a container
      operationId: checkOS
      tags:
        - Src
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
                example: [
                  "latest": 3.7,
                  "version": 3.6,
                  "name": "alpine"
                ]
        '404':
          description: No source code found
          schema:
            type: object
            properties:
              error:
                type: string
                example: No source code found
        '500':
          description: No image name provided
          schema:
            type: object
            properties:
              error:
                type: string
                example: No image name provided'
  /src/tests:
    post:
      summary: Run npm tests
      description: run npm tests inside extracted container's source code
      operationId: runNpmTests
      tags:
        - Src
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              testResults:
                type: string
                example: [
                  "> node ./node_modules/mocha/bin/mocha ./tests/accept/*.js --timeout 15000",
                  "  /GET square",
                  "    ✓ should return correct square result",
                  "  /GET cube",
                  "    ✓ should return correct cube result",
                  "  /GET sum",
                  "    ✓ should return correct sum",
                  "  /GET difference",
                  "    ✓ should return correct subtraction result",
                  "  /GET division",
                  "    ✓ should return correct division result",
                  "  5 passing (29ms)"
                ]
        '404':
          description: No source code found
          schema:
            type: object
            properties:
              error:
                type: string
                example: No source code found
        '500':
          description: No image name provided
          schema:
            type: object
            properties:
              error:
                type: string
                example: No image name provided
...
  /src/update:
    post:
      summary: update npm components
      description: update npm components and reinstall node_modules
      operationId: updateComponents
      tags:
        - Src
      parameters:
        - $ref: '#/components/parameters/token'
        - $ref: '#/components/parameters/imageName'
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              updatedModules:
                type: string
                example:  [
                  "debug          2.6.7  →   3.1.0",
                  "mqtt           2.0.1  →  2.15.3",
                  "express       4.15.3  →  4.16.2",
                  "mathjs        3.16.0  →   4.0.0",
                  "st             1.2.1  →   1.2.2",
                  "coffeescript   1.0.1  →   2.2.2",
                  "swagger-ui     2.2.0  →  3.11.0",
                  "mocha         ^4.1.0  →  ^5.0.1",
                  "sinon  ^4.1.3  →  ^4.4.2"
                ]
        '404':
          description: No source code found
          schema:
            type: object
            properties:
              error:
                type: string
                example: No source code found
        '500':
          description: No image name provided
          schema:
            type: object
            properties:
              error:
                type: string
                example: No image name provided
...